Rename project: packet -> sheriff-buzz
the project has a name :-D
- packet.h (shared definitions between user + BPF) -> bpf_common.h
- default BPF object file now sheriff-buzz.bpf.o
- simply Makefile naming with $(BASE_NAME)

+++ src/include/packet_data.h
@@ -31,10 +31,8 @@
 #define MAX_DB_TASKS 20
 
 struct packet_count {
-    unsigned long val;
-    /* carry when val reaches ULONG_MAX (starts at 0)
-     * TODO: how big should carry be? */
-    int carry;
+    unsigned int val;
+    unsigned int carry; /* carry every 100,000 */
 };
 
 /**
@@ -66,6 +64,7 @@ struct value {
 	GHashTable *ports;
 };
 
+/* port range for writing port-based alerts to the database */
 struct port_range {
 	int min;
 	int max;
